Ref(backend): OccupationGroup module index.test.ts refactor and coverage tests for other uncovered logics
