+++ backend/src/esco/occupationGroup/_shared/testDataHelper.ts
@@ -1,11 +1,6 @@
 import { randomUUID } from "node:crypto";
 import { getMockStringId } from "_test_utilities/mockMongoId";
-import {
-  INewOccupationGroupSpec,
-  IOccupationGroup,
-  IOccupationGroupChild,
-  IOccupationGroupReference,
-} from "./OccupationGroup.types";
+import { INewOccupationGroupSpec, IOccupationGroup, IOccupationGroupChild } from "./OccupationGroup.types";
 import { getRandomString, getTestString } from "_test_utilities/getMockRandomData";
 import OccupationGroupAPISpecs from "api-specifications/esco/occupationGroup";
 import { getMockRandomISCOGroupCode } from "_test_utilities/mockOccupationGroupCode";
@@ -51,36 +46,6 @@ export function getIOccupationGroupMockData(
     parent: null,
     importId: getMockStringId(n),
     children: [],
-    modelId: modelId,
-    createdAt: new Date(1973, 11, 17, 0, 0, 0), //.toISOString(),
-    updatedAt: new Date(),
-  };
-}
-
-export function getIOccupationGroupMockDataWithOccupationGroupChildren(
-  n: number = 1,
-  modelId: string = getMockStringId(1000 + n)
-): IOccupationGroup {
-  const children: IOccupationGroupReference = {
-    UUID: randomUUID(),
-    code: getRandomString(OccupationGroupAPISpecs.Constants.CODE_MAX_LENGTH),
-    id: getMockStringId(n + 1),
-    preferredLabel: getRandomString(OccupationGroupAPISpecs.Constants.PREFERRED_LABEL_MAX_LENGTH),
-    objectType: OccupationGroupAPISpecs.Enums.Relations.Children.ObjectTypes.ISCOGroup,
-  };
-  return {
-    id: getMockStringId(n),
-    UUID: randomUUID(),
-    UUIDHistory: [randomUUID()],
-    code: getRandomString(OccupationGroupAPISpecs.Constants.CODE_MAX_LENGTH),
-    originUri: "https://foo.bar/" + n,
-    preferredLabel: getRandomString(OccupationGroupAPISpecs.Constants.PREFERRED_LABEL_MAX_LENGTH),
-    altLabels: [getRandomString(OccupationGroupAPISpecs.Constants.ALT_LABEL_MAX_LENGTH)],
-    groupType: OccupationGroupAPISpecs.Enums.ObjectTypes.ISCOGroup,
-    description: getRandomString(OccupationGroupAPISpecs.Constants.DESCRIPTION_MAX_LENGTH),
-    parent: null,
-    importId: getMockStringId(n),
-    children: [children],
     modelId: modelId,
     createdAt: new Date(1973, 11, 17, 0, 0, 0), //.toISOString(),
     updatedAt: new Date(),
